text-to-sql-system/backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
pandas==2.1.1
httpx==0.25.0
sqlalchemy==2.0.22
aiofiles==23.2.1
jinja2==3.1.2
numpy==1.24.3
scikit-learn==1.2.2
scipy==1.10.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
google-genai==1.1.0
